Identifying_unique_developers_in_oss/requirements.txt
# Requires Python 3.10 or higher
gitpython>=3.1.45
pandas>=2.3.3
pydriller>=2.5.0
rapidfuzz>=3.5.2
python-Levenshtein>=0.23.0
unidecode>=1.3.7
pytest>=7.4.0
pytest-cov>=5.0.0
pylint>=3.2.0
